# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QUANTEDGE API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /forecasts/{symbol}:
    get:
      operationId: getForecasts
      tags: [forecasts]
      summary: Get all model forecasts for a symbol
      parameters:
        - in: path
          name: symbol
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForecastBundle"
  /forecasts/{symbol}/run:
    post:
      operationId: runForecast
      tags: [forecasts]
      summary: Trigger forecast computation for symbol
      parameters:
        - in: path
          name: symbol
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForecastBundle"
...
  /signals/{id}/explain:
    get:
      operationId: explainSignal
      tags: [signals]
      summary: Get explainability breakdown for a signal
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignalExplanation"

  # ORDERS
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: Get paper trading orders
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [open, executed, rejected, all]
            default: all
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: placeOrder
      tags: [orders]
      summary: Place paper trade order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # MARKET REGIME
...
  /backtests:
    get:
      operationId: getBacktests
      tags: [backtests]
      summary: List saved backtests
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Backtest"
    post:
      operationId: runBacktest
      tags: [backtests]
      summary: Run a new backtest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BacktestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Backtest"
  /backtests/{id}:
    get:
      operationId: getBacktest
      tags: [backtests]
      summary: Get backtest detail with full results
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BacktestDetail"

  # AI COMMENTARY
  /ai/commentary:
    post:
      operationId: getAiCommentary
      tags: [ai]
      summary: Get AI commentary for portfolio or stock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiCommentaryInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiCommentary"
  /ai/signal-explanation:
    post:
      operationId: getAiSignalExplanation
      tags: [ai]
      summary: Get AI natural language explanation for a signal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiSignalInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiCommentary"

  # MARKET DATA
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email: { type: string }
        password: { type: string }
        name: { type: string }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, email, name, createdAt]
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
        createdAt: { type: string }

    # DASHBOARD
...
    ModelForecast:
      type: object
      required: [model, nextDay, nextWeek, nextMonth, confidence, direction, rmse, mae, mape, r2, directionalAccuracy]
      properties:
        model: { type: string }
        nextDay: { type: number }
        nextWeek: { type: number }
        nextMonth: { type: number }
        confidence: { type: number }
        direction: { type: string }
        rmse: { type: number }
        mae: { type: number }
        mape: { type: number }
        r2: { type: number }
        directionalAccuracy: { type: number }
        forecastPoints:
          type: array
          items:
            $ref: "#/components/schemas/ForecastPoint"
    ForecastPoint:
      type: object
      required: [date, value, lower, upper]
      properties:
        date: { type: string }
        value: { type: number }
        lower: { type: number }
        upper: { type: number }
    ForecastBundle:
      type: object
      required: [symbol, currentPrice, arima, lstm, xgboost, ensemble, modelAgreement, regimeAdjusted, updatedAt]
      properties:
        symbol: { type: string }
        currentPrice: { type: number }
        arima:
          $ref: "#/components/schemas/ModelForecast"
        lstm:
          $ref: "#/components/schemas/ModelForecast"
        xgboost:
          $ref: "#/components/schemas/ModelForecast"
        ensemble:
          $ref: "#/components/schemas/ModelForecast"
        modelAgreement: { type: number }
        regimeAdjusted: { type: boolean }
        updatedAt: { type: string }
...
    SignalExplanation:
      type: object
      required: [signalId, symbol, action, arimaContribution, lstmContribution, xgboostContribution, regimeContribution, optimizationContribution, confidenceScore, reasoning]
      properties:
        signalId: { type: integer }
        symbol: { type: string }
        action: { type: string }
        arimaContribution: { type: number }
        lstmContribution: { type: number }
        xgboostContribution: { type: number }
        regimeContribution: { type: number }
        optimizationContribution: { type: number }
        confidenceScore: { type: number }
        reasoning: { type: string }

    # ORDERS
    Order:
      type: object
      required: [id, symbol, company, side, quantity, price, status, signalSource, executedAt, createdAt]
      properties:
        id: { type: integer }
        symbol: { type: string }
        company: { type: string }
        side: { type: string }
        quantity: { type: integer }
        price: { type: number }
        status: { type: string }
        signalSource: { type: string }
        executedAt: { type: ["string", "null"] }
        createdAt: { type: string }
    OrderInput:
      type: object
      required: [symbol, side, quantity, price, signalSource]
      properties:
        symbol: { type: string }
        side: { type: string }
        quantity: { type: integer }
        price: { type: number }
        signalSource: { type: string }

    # REGIME
...
    TimeSeriesPoint:
      type: object
      required: [date, value]
      properties:
        date: { type: string }
        value: { type: number }
...
    Backtest:
      type: object
      required: [id, name, symbols, startDate, endDate, forecastModel, optimizationMethod, status, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        symbols:
          type: array
          items:
            type: string
        startDate: { type: string }
        endDate: { type: string }
        forecastModel: { type: string }
        optimizationMethod: { type: string }
        status: { type: string }
        cagr: { type: ["number", "null"] }
        sharpeRatio: { type: ["number", "null"] }
        maxDrawdown: { type: ["number", "null"] }
        alpha: { type: ["number", "null"] }
        createdAt: { type: string }
    BacktestInput:
      type: object
      required: [name, symbols, startDate, endDate, forecastModel, optimizationMethod]
      properties:
        name: { type: string }
        symbols:
          type: array
          items:
            type: string
        startDate: { type: string }
        endDate: { type: string }
        forecastModel: { type: string }
        optimizationMethod: { type: string }
    BacktestDetail:
      type: object
      required: [id, name, metrics, equityCurve, drawdownCurve, monthlyReturns, tradeLogs, benchmarkComparison]
      properties:
        id: { type: integer }
        name: { type: string }
        metrics:
          $ref: "#/components/schemas/BacktestMetrics"
        equityCurve:
          type: array
          items:
            $ref: "#/components/schemas/TimeSeriesPoint"
        drawdownCurve:
          type: array
          items:
            $ref: "#/components/schemas/TimeSeriesPoint"
        monthlyReturns:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyReturn"
        tradeLogs:
          type: array
          items:
            $ref: "#/components/schemas/TradeLog"
        benchmarkComparison:
          type: array
          items:
            $ref: "#/components/schemas/TimeSeriesPoint"
    BacktestMetrics:
      type: object
      required: [totalReturn, cagr, sharpeRatio, sortinoRatio, maxDrawdown, alpha, beta, informationRatio, trackingError, winRate, profitFactor, totalTrades]
      properties:
        totalReturn: { type: number }
        cagr: { type: number }
        sharpeRatio: { type: number }
        sortinoRatio: { type: number }
        maxDrawdown: { type: number }
        alpha: { type: number }
        beta: { type: number }
        informationRatio: { type: number }
        trackingError: { type: number }
        winRate: { type: number }
        profitFactor: { type: number }
        totalTrades: { type: integer }
    MonthlyReturn:
      type: object
      required: [month, year, return]
      properties:
        month: { type: integer }
        year: { type: integer }
        return: { type: number }
    TradeLog:
      type: object
      required: [date, symbol, side, quantity, price, pnl]
      properties:
        date: { type: string }
        symbol: { type: string }
        side: { type: string }
        quantity: { type: integer }
        price: { type: number }
        pnl: { type: number }

    # AI
    AiCommentaryInput:
      type: object
      required: [type, context]
      properties:
        type: { type: string }
        symbol: { type: ["string", "null"] }
        context: { type: string }
    AiSignalInput:
      type: object
      required: [signalId, symbol, action, metrics]
      properties:
        signalId: { type: integer }
        symbol: { type: string }
        action: { type: string }
        metrics: { type: string }
    AiCommentary:
      type: object
      required: [text, generatedAt]
      properties:
        text: { type: string }
        generatedAt: { type: string }

    # MARKET
...